requirements.txt
django==4.0.3
mysqlclient==2.1.0
django-crispy-forms==1.14.0